+++ Sources/Classes/core/Display/RealMiniAppView.swift
@@ -461,3 +461,18 @@ extension RealMiniAppView: MiniAppManageDelegate {
         self.closeAlertInfo = alertInfo
     }
 }
+
+// MARK: - Universal Bridge
+extension RealMiniAppView {
+    func sendJsonToMiniApp(string jsonString:String) {
+        //Post custom event to send the string content to mini app.
+        NotificationCenter.default.sendCustomEvent(
+            MiniAppEvent.Event(
+                miniAppId: self.miniAppId ?? "",
+                miniAppVersion: self.miniAppVersion ?? "",
+                type: .resume,
+                comment: jsonString
+            )
+        )
+    }
+}